+++ packages/chat-stateful-client/src/StatefulChatClient.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ChatClient, ChatThreadItem } from '@azure/communication-chat';
+import { ChatThreadItem } from '@azure/communication-chat';
 import {
   ChatMessageDeletedEvent,
   ChatMessageEditedEvent,
@@ -15,107 +15,24 @@ import {
   ReadReceiptReceivedEvent,
   TypingIndicatorReceivedEvent
 } from '@azure/communication-signaling';
-import { createStatefulChatClientWithDeps, StatefulChatClient, StatefulChatClientArgs } from './StatefulChatClient';
+import { createStatefulChatClientWithDeps } from './StatefulChatClient';
 import { ChatClientState, ChatError } from './ChatClientState';
 import { Constants } from './Constants';
-import { createMockChatThreadClient } from './mocks/createMockChatThreadClient';
-import { createMockIterator } from './mocks/createMockIterator';
-import { MockCommunicationUserCredential } from './mocks/MockCommunicationUserCredential';
+import {
+  StateChangeListener,
+  StatefulChatClientWithEventTrigger,
+  createMockChatClient,
+  createStatefulChatClientMock,
+  defaultClientArgs,
+  mockChatThreads
+} from './TestHelpers';
 
 jest.useFakeTimers();
-
-// [1, 2 ... 5] array
-const seedArray = Array.from(Array(5).keys());
-
-const mockChatThreads: ChatThreadItem[] = seedArray.map((seed) => ({
-  id: 'chatThreadId' + seed,
-  topic: 'topic' + seed,
-  createdOn: new Date(seed * 10000),
-  createdBy: { communicationUserId: 'user' + seed }
-}));
 
 const mockParticipants: ChatParticipant[] = [
   { id: { kind: 'communicationUser', communicationUserId: 'user1' }, displayName: 'user1' },
   { id: { kind: 'communicationUser', communicationUserId: 'user2' }, displayName: 'user1' }
 ];
-
-const mockListChatThreads = (): any => {
-  return createMockIterator(mockChatThreads);
-};
-
-const emptyAsyncFunctionWithResponse = async (): Promise<any> => {
-  return { _response: {} as any };
-};
-
-const mockEventHandlersRef = { value: {} };
-beforeEach(() => {
-  mockEventHandlersRef.value = {};
-});
-
-export function createMockChatClient(): ChatClient {
-  const mockChatClient: ChatClient = {} as any;
-
-  mockChatClient.createChatThread = async (request) => {
-    return {
-      chatThread: {
-        id: 'chatThreadId',
-        topic: request.topic,
-        createdOn: new Date(0),
-        createdBy: { kind: 'communicationUser', communicationUserId: 'user1' }
-      }
-    };
-  };
-
-  mockChatClient.listChatThreads = mockListChatThreads;
-
-  mockChatClient.deleteChatThread = emptyAsyncFunctionWithResponse;
-
-  mockChatClient.getChatThreadClient = (threadId) => {
-    return createMockChatThreadClient(threadId);
-  };
-
-  mockChatClient.on = ((event: Parameters<ChatClient['on']>[0], listener: (e: Event) => void) => {
-    mockEventHandlersRef.value[event] = listener;
-  }) as any;
-
-  mockChatClient.off = ((event: Parameters<ChatClient['on']>[0], listener: (e: Event) => void) => {
-    if (mockEventHandlersRef.value[event] === listener) {
-      mockEventHandlersRef.value[event] = undefined;
-    }
-  }) as any;
-
-  mockChatClient.startRealtimeNotifications = emptyAsyncFunctionWithResponse;
-  mockChatClient.stopRealtimeNotifications = emptyAsyncFunctionWithResponse;
-
-  return mockChatClient;
-}
-
-type StatefulChatClientWithEventTrigger = StatefulChatClient & {
-  triggerEvent: (eventName: string, e: any) => Promise<void>;
-};
-
-const createStatefulChatClientMock = (): StatefulChatClientWithEventTrigger => {
-  mockEventHandlersRef.value = {};
-  const client = createStatefulChatClientWithDeps(createMockChatClient(), defaultClientArgs);
-
-  Object.defineProperty(client, 'triggerEvent', {
-    value: async (eventName: string, e: any): Promise<void> => {
-      const handler = mockEventHandlersRef.value[eventName];
-      if (handler !== undefined) {
-        await handler(e);
-      }
-    }
-  });
-
-  return client as StatefulChatClientWithEventTrigger;
-};
-
-export const defaultClientArgs: StatefulChatClientArgs = {
-  displayName: '',
-  userId: { kind: 'communicationUser', communicationUserId: 'userId1' },
-  endpoint: '',
-  credential: new MockCommunicationUserCredential()
-};
 
 describe('declarative chatThread list iterators', () => {
   test('declarative listChatThreads should proxy listChatThreads iterator and store it in internal state', async () => {
@@ -449,6 +366,19 @@ describe('stateful chatClient tees errors to state', () => {
     const latestError = listener.state.latestErrors['ChatClient.startRealtimeNotifications'];
     expect(latestError).toBeDefined();
   });
+
+  test('when startRealtimeNotifications fails', async () => {
+    const baseClient = createMockChatClient();
+    baseClient.stopRealtimeNotifications = async () => {
+      throw Error('injected error');
+    };
+    const client = createStatefulChatClientWithDeps(baseClient, defaultClientArgs);
+    const listener = new StateChangeListener(client);
+    await expect(client.stopRealtimeNotifications()).rejects.toThrow();
+    expect(listener.onChangeCalledCount).toBe(1);
+    const latestError = listener.state.latestErrors['ChatClient.stopRealtimeNotifications'];
+    expect(latestError).toBeDefined();
+  });
 });
 
 describe('complex error handling for startRealtimeNotifications', () => {
@@ -513,18 +443,3 @@ describe('complex error handling for startRealtimeNotifications', () => {
     expect(latestError).toBeUndefined();
   });
 });
-
-export class StateChangeListener {
-  state: ChatClientState;
-  onChangeCalledCount = 0;
-
-  constructor(client: StatefulChatClient) {
-    this.state = client.getState();
-    client.onStateChange(this.onChange.bind(this));
-  }
-
-  private onChange(newState: ChatClientState): void {
-    this.onChangeCalledCount++;
-    this.state = newState;
-  }
-}
